comments out all sorting related props, functions, icons and tests
